Measure 2: move the pipeline process wrapper to node
ServiceManagerProcess lived in datapipeline but extends node's
AbstractProcess and implements node's IServiceProvider — its own JavaDoc
opens by calling it a Node process. It is a pure lifecycle wrapper, so it
belongs on node's side, where the dependency runs in the intended
direction.

Target package follows the existing convention: every concrete process
sits in a sub-package named after the domain it wraps (broker, h2, http),
and "pipeline" is already the name the process is configured under.

The class itself is unchanged apart from its package, an import of
ServiceManager that was implicit while both shared a package, and the
class name in its own configuration example. Node instantiates processes
reflectively, so reference.conf and NodeIntegrationTest carry the new
name too; nothing else in the repository names the old one.

No compatibility alias: an outside configuration using the old name fails
at startup with ClassNotFoundException, which is what a misconfiguration
should do.

The datapipeline rule now reports 10 violations instead of 13, all from
the two Artemis resources that measure 3 addresses.
