+++ src/main/resources/reference.conf
@@ -2452,7 +2452,7 @@ node {
     # Pipeline process — manages data pipeline services (simulation, persistence, etc.)
     # Depends on both brokers to guarantee they are running before resources connect.
     pipeline {
-      className = "org.evochora.datapipeline.ServiceManagerProcess"
+      className = "org.evochora.node.processes.pipeline.ServiceManagerProcess"
       require = {
         topicBroker = "topic-broker"
         queueBroker = "queue-broker"
